added bugs.txt for bugtracking, and stylechanges
