+++ frontend/src/components/mdTextEditor/settingsFiles/buttonConfig.js
@@ -232,80 +232,80 @@ const sections = [
   {
     bTitle: "sec_activity",
     icon: "",
-    output: "# " + SECTION_TEXT + " {.activity}\n\n",
+    output: "# " + SECTION_TEXT + " {.activity}\n",
     title: BUTTON_TITLE.activity,
     cursorIntON: 0,
-    cursorIntOFF: 14,
+    cursorIntOFF: 13,
     endOutput: "",
     shortcut: KEY_COMBINATIONS.activity.join(""),
   },
   {
     bTitle: "sec_intro",
     icon: "",
-    output: "# " + SECTION_TEXT + " {.intro}\n\n",
+    output: "# " + SECTION_TEXT + " {.intro}\n",
     title: BUTTON_TITLE.intro,
     cursorIntON: 0,
-    cursorIntOFF: 11,
+    cursorIntOFF: 10,
     endOutput: "",
     shortcut: KEY_COMBINATIONS.intro.join(""),
   },
   {
     bTitle: "sec_check",
     icon: "",
-    output: "## " + SECTION_TEXT + " {.check}\n\n",
+    output: "## " + SECTION_TEXT + " {.check}\n",
     title: BUTTON_TITLE.check,
     cursorIntON: 0,
-    cursorIntOFF: 11,
+    cursorIntOFF: 10,
     endOutput: "",
     shortcut: KEY_COMBINATIONS.check.join(""),
   },
   {
     bTitle: "sec_tip",
     icon: "",
-    output: "## " + SECTION_TEXT + " {.tip}\n\n",
+    output: "## " + SECTION_TEXT + " {.tip}\n",
     title: BUTTON_TITLE.tip,
     cursorIntON: 0,
-    cursorIntOFF: 9,
+    cursorIntOFF: 8,
     endOutput: "",
     shortcut: KEY_COMBINATIONS.tip.join(""),
   },
   {
     bTitle: "sec_protip",
     icon: "",
-    output: "## " + SECTION_TEXT + " {.protip}\n\n",
+    output: "## " + SECTION_TEXT + " {.protip}\n",
     title: BUTTON_TITLE.protip,
     cursorIntON: 0,
-    cursorIntOFF: 12,
+    cursorIntOFF: 11,
     endOutput: "",
     shortcut: KEY_COMBINATIONS.protip.join(""),
   },
   {
     bTitle: "sec_challenge",
     icon: "",
-    output: "## " + SECTION_TEXT + " {.challenge}\n\n",
+    output: "## " + SECTION_TEXT + " {.challenge}\n",
     title: BUTTON_TITLE.challenge,
     cursorIntON: 0,
-    cursorIntOFF: 15,
+    cursorIntOFF: 14,
     endOutput: "",
     shortcut: KEY_COMBINATIONS.challenge.join(""),
   },
   {
     bTitle: "sec_flag",
     icon: "",
-    output: "## " + SECTION_TEXT + " {.flag}\n\n",
+    output: "## " + SECTION_TEXT + " {.flag}\n",
     title: BUTTON_TITLE.flag,
     cursorIntON: 0,
-    cursorIntOFF: 10,
+    cursorIntOFF: 9,
     endOutput: "",
     shortcut: KEY_COMBINATIONS.flag.join(""),
   },
   {
     bTitle: "sec_try",
     icon: "",
-    output: "# " + SECTION_TEXT + " {.try}\n\n",
+    output: "# " + SECTION_TEXT + " {.try}\n",
     title: BUTTON_TITLE.try,
     cursorIntON: 0,
-    cursorIntOFF: 9,
+    cursorIntOFF: 8,
     endOutput: "",
     shortcut: KEY_COMBINATIONS.try.join(""),
   },
